move propType test to a new file
